8330005: RandomGeneratorFactory.getDefault() throws exception when the runtime image only has java.base module
