+++ src/java.base/share/classes/java/util/random/RandomGeneratorFactory.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022, 2022, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2022, 2024, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -142,8 +142,9 @@ public final class RandomGeneratorFactory<T extends RandomGenerator> {
          * @return Map of RandomGeneratorFactory classes.
          */
         private static Map<String, Provider<? extends RandomGenerator>> createFactoryMap() {
+            FactoryMapHolder.class.getModule().addUses(RandomGenerator.class);
             return ServiceLoader
-                .load(RandomGenerator.class)
+                .load(RandomGenerator.class, ClassLoader.getPlatformClassLoader())
                 .stream()
                 .filter(p -> !p.type().isInterface())
                 .collect(Collectors.toMap(p -> p.type().getSimpleName(), Function.identity()));
@@ -341,9 +342,6 @@ public final class RandomGeneratorFactory<T extends RandomGenerator> {
      * {@link RandomGenerator} that utilize the {@code name}
      * <a href="package-summary.html#algorithms">algorithm</a>.
      *
-     * @implSpec Availability is determined by RandomGeneratorFactory using the
-     * service provider API to locate implementations of the RandomGenerator interface.
-     *
      * @param name  Name of random number generator
      * <a href="package-summary.html#algorithms">algorithm</a>
      * @param <T> Sub-interface of {@link RandomGenerator} to produce
@@ -376,11 +374,8 @@ public final class RandomGeneratorFactory<T extends RandomGenerator> {
 
     /**
      * Returns a non-empty stream of available {@link RandomGeneratorFactory RandomGeneratorFactory(s)}.
-     *
+     * <p>
      * RandomGenerators that are marked as deprecated are not included in the result.
-     *
-     * @implSpec Availability is determined by RandomGeneratorFactory using the service provider API
-     * to locate implementations of the RandomGenerator interface.
      *
      * @return a non-empty stream of all available {@link RandomGeneratorFactory RandomGeneratorFactory(s)}.
      */
@@ -615,5 +610,3 @@ public final class RandomGeneratorFactory<T extends RandomGenerator> {
     }
 
 }
-
-
