+++ src/java.base/share/classes/jdk/internal/random/L64X128MixRandom.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2021, 2024, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -23,7 +23,7 @@
  * questions.
  */
 
-package jdk.random;
+package jdk.internal.random;
 
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.random.RandomGenerator;
@@ -33,39 +33,39 @@ import jdk.internal.util.random.RandomSupport.RandomGeneratorProperties;
 
 /**
  * A "splittable" pseudorandom number generator (PRNG) whose period
- * is roughly 2<sup>96</sup>.  Class {@link L32X64MixRandom} implements
+ * is roughly 2<sup>192</sup>.  Class {@link L64X128MixRandom} implements
  * interfaces {@link RandomGenerator} and {@link SplittableGenerator},
  * and therefore supports methods for producing pseudorandomly chosen
  * values of type {@code int}, {@code long}, {@code float}, {@code double},
  * and {@code boolean} (and for producing streams of pseudorandomly chosen
  * numbers of type {@code int}, {@code long}, and {@code double}),
- * as well as methods for creating new split-off {@link L32X64MixRandom}
+ * as well as methods for creating new split-off {@link L64X128MixRandom}
  * objects or streams of such objects.
  *
- * <p>The {@link L32X64MixRandom} algorithm is a specific member of
+ * <p>The {@link L64X128MixRandom} algorithm is a specific member of
  * the LXM family of algorithms for pseudorandom number generators;
  * for more information, see the documentation for package
- * {@link jdk.random}.  Each instance of {@link L32X64MixRandom}
- * has 96 bits of state plus one 32-bit instance-specific parameter.
+ * {@link jdk.random}.  Each instance of {@link L64X128MixRandom}
+ * has 192 bits of state plus one 64-bit instance-specific parameter.
  *
- * <p>If two instances of {@link L32X64MixRandom} are created with
+ * <p>If two instances of {@link L64X128MixRandom} are created with
  * the same seed within the same program execution, and the same
  * sequence of method calls is made for each, they will generate and
  * return identical sequences of values.
  *
  * <p>As with {@link java.util.SplittableRandom}, instances of
- * {@link L32X64MixRandom} are <em>not</em> thread-safe.  They are
+ * {@link L64X128MixRandom} are <em>not</em> thread-safe.  They are
  * designed to be split, not shared, across threads (see the {@link #split}
  * method). For example, a {@link java.util.concurrent.ForkJoinTask}
  * fork/join-style computation using random numbers might include a
  * construction of the form
- * {@code new Subtask(someL32X64MixRandom.split()).fork()}.
+ * {@code new Subtask(someL64X128MixRandom.split()).fork()}.
  *
  * <p>This class provides additional methods for generating random
  * streams, that employ the above techniques when used in
  * {@code stream.parallel()} mode.
  *
- * <p>Instances of {@link L32X64MixRandom} are not cryptographically
+ * <p>Instances of {@link L64X128MixRandom} are not cryptographically
  * secure.  Consider instead using {@link java.security.SecureRandom}
  * in security-sensitive applications. Additionally,
  * default-constructed instances do not use a cryptographically random
@@ -76,26 +76,27 @@ import jdk.internal.util.random.RandomSupport.RandomGeneratorProperties;
  *
  */
 @RandomGeneratorProperties(
-        name = "L32X64MixRandom",
+        name = "L64X128MixRandom",
         group = "LXM",
-        i = 64, j = 1, k = 32,
-        equidistribution = 1
+        i = 128, j = 1, k = 64,
+        equidistribution = 2
 )
-public final class L32X64MixRandom extends AbstractSplittableWithBrineGenerator {
+public final class L64X128MixRandom extends AbstractSplittableWithBrineGenerator {
+
     /*
      * Implementation Overview.
      *
-     * The split operation uses the current generator to choose four new 32-bit
-     * int values that are then used to initialize the parameter `a` and the
+     * The split operation uses the current generator to choose four new 64-bit
+     * long values that are then used to initialize the parameter `a` and the
      * state variables `s`, `x0`, and `x1` for a newly constructed generator.
      *
-     * With high probability, no two generators so chosen will have the same
-     * `a` parameter, and testing has indicated that the values generated by
-     * two instances of {@link L32X64MixRandom} will be (approximately)
-     * independent if the two instances have different values for `a`.
+     * With extremely high probability, no two generators so chosen
+     * will have the same `a` parameter, and testing has indicated
+     * that the values generated by two instances of {@link L64X128MixRandom}
+     * will be (approximately) independent if have different values for `a`.
      *
      * The default (no-argument) constructor, in essence, uses
-     * "defaultGen" to generate four new 32-bit values for the same
+     * "defaultGen" to generate four new 64-bit values for the same
      * purpose.  Multiple generators created in this way will certainly
      * differ in their `a` parameters.  The defaultGen state must be accessed
      * in a thread-safe manner, so we use an AtomicLong to represent
@@ -119,11 +120,11 @@ public final class L32X64MixRandom extends AbstractSplittableWithBrineGenerator 
     /*
      * Multiplier used in the LCG portion of the algorithm.
      * Chosen based on research by Sebastiano Vigna and Guy Steele (2019).
-     * The spectral scores for dimensions 2 through 8 for the multiplier 0xadb4a92d
-     * are [0.975884, 0.936244, 0.755793, 0.877642, 0.751300, 0.789333, 0.728869].
+     * The spectral scores for dimensions 2 through 8 for the multiplier 0xd1342543de82ef95L
+     * are [0.958602, 0.937479, 0.870757, 0.822326, 0.820405, 0.813065, 0.760215].
      */
 
-    private static final int M = 0xadb4a92d;
+    private static final long M = 0xd1342543de82ef95L;
 
     /* ---------------- instance fields ---------------- */
 
@@ -131,13 +132,13 @@ public final class L32X64MixRandom extends AbstractSplittableWithBrineGenerator 
      * The parameter that is used as an additive constant for the LCG.
      * Must be odd.
      */
-    private final int a;
+    private final long a;
 
     /**
      * The per-instance state: s for the LCG; x0 and x1 for the XBG.
      * At least one of x0 and x1 must be nonzero.
      */
-    private int s, x0, x1;
+    private long s, x0, x1;
 
     /* ---------------- constructors ---------------- */
 
@@ -151,7 +152,7 @@ public final class L32X64MixRandom extends AbstractSplittableWithBrineGenerator 
      * @param x0 first word of the initial state for the XBG
      * @param x1 second word of the initial state for the XBG
      */
-    public L32X64MixRandom(int a, int s, int x0, int x1) {
+    public L64X128MixRandom(long a, long s, long x0, long x1) {
         // Force a to be odd.
         this.a = a | 1;
         this.s = s;
@@ -159,59 +160,59 @@ public final class L32X64MixRandom extends AbstractSplittableWithBrineGenerator 
         this.x1 = x1;
         // If x0 and x1 are both zero, we must choose nonzero values.
         if ((x0 | x1) == 0) {
-       int v = s;
+       long v = s;
             // At least one of the two values generated here will be nonzero.
-            this.x0 = RandomSupport.mixMurmur32(v += RandomSupport.GOLDEN_RATIO_32);
-            this.x1 = RandomSupport.mixMurmur32(v + RandomSupport.GOLDEN_RATIO_32);
+            this.x0 = RandomSupport.mixStafford13(v += RandomSupport.GOLDEN_RATIO_64);
+            this.x1 = RandomSupport.mixStafford13(v + RandomSupport.GOLDEN_RATIO_64);
         }
     }
 
     /**
-     * Creates a new instance of {@link L32X64MixRandom} using the
+     * Creates a new instance of {@link L64X128MixRandom} using the
      * specified {@code long} value as the initial seed. Instances of
-     * {@link L32X64MixRandom} created with the same seed in the same
+     * {@link L64X128MixRandom} created with the same seed in the same
      * program generate identical sequences of values.
      *
      * @param seed the initial seed
      */
-    public L32X64MixRandom(long seed) {
+    public L64X128MixRandom(long seed) {
         // Using a value with irregularly spaced 1-bits to xor the seed
         // argument tends to improve "pedestrian" seeds such as 0 or
         // other small integers.  We may as well use SILVER_RATIO_64.
         //
-        // The high half of the seed is hashed by mixMurmur32 to produce the `a` parameter.
-        // The low half of the seed is hashed by mixLea32 to produce the initial `x0`,
+        // The seed is hashed by mixMurmur64 to produce the `a` parameter.
+        // The seed is hashed by mixStafford13 to produce the initial `x0`,
         // which will then be used to produce the first generated value.
         // Then x1 is filled in as if by a SplitMix PRNG with
-        // GOLDEN_RATIO_32 as the gamma value and mixLea32 as the mixer.
-        this(RandomSupport.mixMurmur32((int)((seed ^= RandomSupport.SILVER_RATIO_64) >>> 32)),
+        // GOLDEN_RATIO_64 as the gamma value and mixStafford13 as the mixer.
+        this(RandomSupport.mixMurmur64(seed ^= RandomSupport.SILVER_RATIO_64),
              1,
-             RandomSupport.mixLea32((int)(seed)),
-             RandomSupport.mixLea32((int)(seed) + RandomSupport.GOLDEN_RATIO_32));
+             RandomSupport.mixStafford13(seed),
+             RandomSupport.mixStafford13(seed + RandomSupport.GOLDEN_RATIO_64));
     }
 
     /**
-     * Creates a new instance of {@link L32X64MixRandom} that is likely to
+     * Creates a new instance of {@link L64X128MixRandom} that is likely to
      * generate sequences of values that are statistically independent
      * of those of any other instances in the current program execution,
      * but may, and typically does, vary across program invocations.
      */
-    public L32X64MixRandom() {
+    public L64X128MixRandom() {
         // Using GOLDEN_RATIO_64 here gives us a good Weyl sequence of values.
         this(defaultGen.getAndAdd(RandomSupport.GOLDEN_RATIO_64));
     }
 
     /**
-     * Creates a new instance of {@link L32X64MixRandom} using the specified array of
-     * initial seed bytes. Instances of {@link L32X64MixRandom} created with the same
+     * Creates a new instance of {@link L64X128MixRandom} using the specified array of
+     * initial seed bytes. Instances of {@link L64X128MixRandom} created with the same
      * seed array in the same program execution generate identical sequences of values.
      *
      * @param seed the initial seed
      */
-    public L32X64MixRandom(byte[] seed) {
-        // Convert the seed to 4 int values, of which the last 2 are not all zero.
-        int[] data = RandomSupport.convertSeedBytesToInts(seed, 4, 2);
-        int a = data[0], s = data[1], x0 = data[2], x1 = data[3];
+    public L64X128MixRandom(byte[] seed) {
+        // Convert the seed to 4 long values, of which the last 2 are not all zero.
+        long[] data = RandomSupport.convertSeedBytesToLongs(seed, 4, 2);
+        long a = data[0], s = data[1], x0 = data[2], x1 = data[3];
         // Force a to be odd.
         this.a = a | 1;
         this.s = s;
@@ -223,36 +224,31 @@ public final class L32X64MixRandom extends AbstractSplittableWithBrineGenerator 
 
     @Override
     public SplittableGenerator split(SplittableGenerator source, long brine) {
-       // Pick a new instance "at random", but use (the low 31 bits of) the brine for `a`.
-        return new L32X64MixRandom((int)brine << 1, source.nextInt(),
-                   source.nextInt(), source.nextInt());
+       // Pick a new instance "at random", but use the brine for `a`.
+        return new L64X128MixRandom(brine << 1, source.nextLong(),
+                   source.nextLong(), source.nextLong());
     }
 
     @Override
-    public int nextInt() {
+    public long nextLong() {
        // Compute the result based on current state information
        // (this allows the computation to be overlapped with state update).
-        final int result = RandomSupport.mixLea32(s + x0);
+        final long result = RandomSupport.mixLea64(s + x0);
 
        // Update the LCG subgenerator
         s = M * s + a;
 
        // Update the XBG subgenerator
-        int q0 = x0, q1 = x1;
-        {   // xoroshiro64
+        long q0 = x0, q1 = x1;
+        {   // xoroshiro128v1_0
             q1 ^= q0;
-            q0 = Integer.rotateLeft(q0, 26);
-            q0 = q0 ^ q1 ^ (q1 << 9);
-            q1 = Integer.rotateLeft(q1, 13);
+            q0 = Long.rotateLeft(q0, 24);
+            q0 = q0 ^ q1 ^ (q1 << 16);
+            q1 = Long.rotateLeft(q1, 37);
         }
         x0 = q0; x1 = q1;
 
         return result;
-    }
-
-    @Override
-    public long nextLong() {
-        return ((long)nextInt() << 32) ^ (long)nextInt();
     }
 
 }
