+++ src/java.base/share/classes/jdk/internal/random/L64X128StarStarRandom.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2021, 2024, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -23,7 +23,7 @@
  * questions.
  */
 
-package jdk.random;
+package jdk.internal.random;
 
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.random.RandomGenerator;
